refactor(arch): split symbols/handlers.rs into per-tool modules (PR-G) (#325)
After PR-F broke the symbol_query ↔ symbols::handlers cycle, the
remaining non-pipeline tools (bm25_symbol_search, search_symbols_fuzzy,
refresh_symbol_index, get_complexity, get_project_structure) still
shared a single 448-line handlers.rs. PR-G splits them by responsibility:

- `symbols/bm25_search.rs` (277 LOC): bm25_symbol_search +
  file-private suggested_follow_up + confidence_tier + their tests.
- `symbols/fuzzy_search.rs` (61 LOC): search_symbols_fuzzy alone —
  hybrid-with-semantic lane stays distinct from BM25 lane.
- `symbols/inventory.rs` (132 LOC): refresh_symbol_index +
  get_complexity + get_project_structure — the three tools that
  probe/refresh on-disk index state (SymbolIndex / graph_cache /
  embedding_ref triplet).
- `symbols/handlers.rs` (31 LOC, was 448, **-93%**): the three
  pipeline tool stubs (get_symbols_overview, find_symbol,
  get_ranked_context). Each is 3 lines forwarding to
  SymbolQueryPipeline::run.

Dependency direction (post-PR-F) still flows one way:
`symbols::*` → `symbol_query::*`. No new seams reach upward.

Validation:
- cargo clippy --workspace --features http,semantic -- -D warnings: clean
- cargo clippy --workspace --no-default-features -- -D warnings: clean
- cargo clippy --workspace --features scip-backend -- -D warnings: clean
- cargo test -p codelens-mcp --bin codelens-mcp: 593 passed
- cargo fmt --all -- --check: clean
