+++ crates/codelens-mcp/src/tools/symbols/bm25_search.rs
@@ -1,0 +1,277 @@
+//! `bm25_symbol_search` — sparse BM25F lane over the symbol corpus.
+//!
+//! Distinct from the pipeline (`SymbolQueryPipeline`): this is the
+//! single-lane fallback exposed as a top-level tool when the caller
+//! explicitly wants BM25 evidence (lexical match transparency,
+//! debugging, harness probes that don't want hybrid fusion).
+//!
+//! Internal helpers `suggested_follow_up` and `confidence_tier` shape
+//! the per-card metadata the harness consumes. Both stay file-private.
+
+use super::super::{
+    AppState, ToolResult, optional_bool, optional_usize, query_analysis::analyze_retrieval_query,
+    required_string, success_meta,
+};
+use crate::protocol::BackendKind;
+use crate::symbol_retrieval::unique_query_terms;
+use crate::tools::symbol_query::sparse_retriever::sparse_symbol_hits_for_query;
+use serde_json::{Value, json};
+
+pub fn bm25_symbol_search(state: &AppState, arguments: &Value) -> ToolResult {
+    let query = required_string(arguments, "query")?;
+    let query_analysis = analyze_retrieval_query(query);
+    let max_results = optional_usize(arguments, "max_results", 10);
+    let include_tests = optional_bool(arguments, "include_tests", false);
+    let include_generated = optional_bool(arguments, "include_generated", false);
+    let session = crate::session_context::SessionRequestContext::from_json(arguments);
+    let scored = sparse_symbol_hits_for_query(
+        state,
+        &query_analysis,
+        max_results,
+        include_tests,
+        include_generated,
+        &session,
+    )?;
+
+    let total_query_terms = unique_query_terms(&query_analysis.expanded_query).len();
+    let payload_results: Vec<Value> = scored
+        .into_iter()
+        .enumerate()
+        .map(|(idx, hit)| {
+            let follow_up = suggested_follow_up(&hit.document.kind, hit.document.exported);
+            let confidence = confidence_tier(
+                &hit.matched_terms,
+                total_query_terms,
+                &hit.document.name,
+                &hit.document.name_path,
+            );
+            json!({
+                "symbol_id": hit.document.symbol_id,
+                "name": hit.document.name,
+                "name_path": hit.document.name_path,
+                "kind": hit.document.kind,
+                "file_path": hit.document.file_path,
+                "module_path": hit.document.module_path,
+                "signature": hit.document.signature,
+                "language": hit.document.language,
+                "line": hit.document.line_start,
+                "score": ((hit.score * 1000.0).round() / 1000.0),
+                "why_matched": hit.matched_terms,
+                "flags": {
+                    "is_test": hit.document.is_test,
+                    "is_generated": hit.document.is_generated,
+                    "exported": hit.document.exported,
+                },
+                "provenance": {
+                    "source": "sparse_bm25f",
+                    "retrieval_rank": idx + 1,
+                },
+                "suggested_follow_up": follow_up,
+                "confidence": confidence,
+            })
+        })
+        .collect();
+
+    let query_type = if query_analysis.prefer_lexical_only {
+        "identifier"
+    } else if query_analysis.natural_language {
+        "natural_language"
+    } else {
+        "short_phrase"
+    };
+    let retrieval = json!({
+        "lane": "sparse_bm25f",
+        "query_type": query_type,
+        "recommended": query_analysis.prefer_sparse_symbol_search,
+        "lexical_query": query_analysis.expanded_query,
+        "semantic_query": query_analysis.semantic_query,
+    });
+    let meta = success_meta(BackendKind::Sqlite, 0.88);
+    let evidence = crate::tool_evidence::tool_evidence(
+        "retrieval",
+        &meta,
+        "sparse_bm25f",
+        json!({
+            "preferred_lane": "sparse_bm25f",
+            "query_type": query_type,
+            "semantic_enabled": false,
+            "semantic_used_in_core": false,
+            "sparse_used_in_core": true,
+            "semantic_evidence_count": 0,
+            "sparse_evidence_count": payload_results.len(),
+            "precise_available": false,
+            "precise_used": false,
+            "precise_source": null,
+            "fallback_source": "sparse_bm25f",
+            "precise_result_count": 0,
+        }),
+    );
+
+    Ok((
+        json!({
+            "query": query,
+            "results": payload_results,
+            "count": payload_results.len(),
+            "retrieval": retrieval,
+            "evidence": evidence,
+        }),
+        meta,
+    ))
+}
+
+/// Follow-up tool hints for a BM25 symbol card.
+///
+/// Mirrors the `bm25-sparse-lane-spec` matrix. Frontier-model harnesses
+/// select their next tool off this list, so the output is part of the
+/// response contract. Keep it short (1-3 entries) — the goal is
+/// guidance, not an exhaustive menu.
+fn suggested_follow_up(kind: &str, exported: bool) -> Vec<&'static str> {
+    let base: Vec<&'static str> = match kind {
+        "function" | "method" => vec!["find_symbol", "get_file_diagnostics"],
+        "class" | "interface" | "enum" | "type_alias" => {
+            vec!["find_symbol", "find_referencing_symbols"]
+        }
+        "module" | "file" => vec!["get_symbols_overview", "find_referencing_symbols"],
+        "variable" | "property" => vec!["find_symbol", "find_referencing_symbols"],
+        _ => vec!["find_symbol"],
+    };
+    if exported
+        && matches!(kind, "function" | "method" | "class" | "interface")
+        && !base.contains(&"find_referencing_symbols")
+    {
+        let mut with_refs = base.clone();
+        with_refs.push("find_referencing_symbols");
+        return with_refs;
+    }
+    base
+}
+
+/// Cross-field confidence tier for a BM25 symbol card.
+///
+/// Without a separate dense arm, we cannot yet compute a true
+/// BM25-vs-dense agreement signal. This heuristic is the *cross-field*
+/// proxy: a result that matches query terms on the high-weight
+/// identifier fields (`name`, `name_path`) **and** covers most of the
+/// unique query terms is a high-confidence hit; a result that matches
+/// only on low-weight fields (body lexical chunk, doc comment) is low.
+///
+/// - `high`   — ≥80% query-term coverage AND a hit on name or name_path
+/// - `medium` — 2+ matched terms OR a name/name_path hit
+/// - `low`    — single term hit, or matches only on body/doc fields
+fn confidence_tier(
+    matched_terms: &[String],
+    unique_query_terms: usize,
+    name: &str,
+    name_path: &str,
+) -> &'static str {
+    if matched_terms.is_empty() || unique_query_terms == 0 {
+        return "low";
+    }
+    let coverage = matched_terms.len() as f64 / unique_query_terms as f64;
+    let name_lower = name.to_ascii_lowercase();
+    let name_path_lower = name_path.to_ascii_lowercase();
+    let identifier_hit = matched_terms.iter().any(|term| {
+        let term_lower = term.to_ascii_lowercase();
+        name_lower.contains(&term_lower) || name_path_lower.contains(&term_lower)
+    });
+
+    if coverage >= 0.8 && identifier_hit {
+        "high"
+    } else if identifier_hit || matched_terms.len() >= 2 {
+        "medium"
+    } else {
+        "low"
+    }
+}
+
+#[cfg(test)]
+mod suggested_follow_up_tests {
+    use super::suggested_follow_up;
+
+    #[test]
+    fn function_gets_body_then_diagnostics() {
+        let hints = suggested_follow_up("function", false);
+        assert_eq!(hints.first().copied(), Some("find_symbol"));
+        assert!(hints.contains(&"get_file_diagnostics"));
+    }
+
+    #[test]
+    fn class_gets_body_and_references() {
+        let hints = suggested_follow_up("class", false);
+        assert_eq!(hints, vec!["find_symbol", "find_referencing_symbols"]);
+    }
+
+    #[test]
+    fn module_gets_overview_first() {
+        let hints = suggested_follow_up("module", false);
+        assert_eq!(hints.first().copied(), Some("get_symbols_overview"));
+    }
+
+    #[test]
+    fn exported_function_also_offers_references() {
+        let hints = suggested_follow_up("function", true);
+        assert!(hints.contains(&"find_referencing_symbols"));
+        assert!(hints.contains(&"find_symbol"));
+    }
+
+    #[test]
+    fn unknown_kind_falls_back_to_find_symbol() {
+        let hints = suggested_follow_up("unknown", false);
+        assert_eq!(hints, vec!["find_symbol"]);
+    }
+}
+
+#[cfg(test)]
+mod confidence_tier_tests {
+    use super::confidence_tier;
+
+    #[test]
+    fn full_coverage_on_name_path_is_high() {
+        let matched = vec!["dispatch".to_owned(), "tool".to_owned()];
+        assert_eq!(
+            confidence_tier(&matched, 2, "dispatch_tool", "dispatch::dispatch_tool"),
+            "high"
+        );
+    }
+
+    #[test]
+    fn partial_coverage_with_name_hit_is_medium() {
+        let matched = vec!["dispatch".to_owned()];
+        assert_eq!(
+            confidence_tier(&matched, 3, "dispatch_tool", "dispatch::dispatch_tool"),
+            "medium"
+        );
+    }
+
+    #[test]
+    fn body_only_match_is_low() {
+        let matched = vec!["invoke".to_owned()];
+        assert_eq!(
+            confidence_tier(&matched, 2, "dispatch_tool", "dispatch::dispatch_tool"),
+            "low"
+        );
+    }
+
+    #[test]
+    fn multiple_matches_without_name_hit_is_medium() {
+        let matched = vec!["invoke".to_owned(), "handler".to_owned()];
+        assert_eq!(
+            confidence_tier(&matched, 3, "dispatch_tool", "dispatch::dispatch_tool"),
+            "medium"
+        );
+    }
+
+    #[test]
+    fn empty_matched_is_low() {
+        assert_eq!(confidence_tier(&[], 2, "x", "a::x"), "low");
+    }
+
+    #[test]
+    fn zero_query_terms_is_low() {
+        let matched = vec!["dispatch".to_owned()];
+        assert_eq!(
+            confidence_tier(&matched, 0, "dispatch_tool", "dispatch::dispatch_tool"),
+            "low"
+        );
+    }
+}
