+++ crates/codelens-mcp/src/tools/symbols/handlers.rs
@@ -1,448 +1,31 @@
-use super::super::{
-    AppState, ToolResult, optional_bool, optional_string, optional_usize,
-    query_analysis::analyze_retrieval_query, required_string, success_meta,
-};
-use super::{analyzer::semantic_scores_for_query, formatter::count_branches};
-use crate::protocol::BackendKind;
-use crate::symbol_retrieval::unique_query_terms;
-use crate::tools::symbol_query::sparse_retriever::{flatten_symbols, sparse_symbol_hits_for_query};
+//! Pipeline tool entries.
+//!
+//! `get_symbols_overview`, `find_symbol`, and `get_ranked_context`
+//! all dispatch through `SymbolQueryPipeline` — see
+//! `crate::tools::symbol_query` for the deep module. Each entry here
+//! is intentionally a 3-line stub; the orchestration body lives
+//! inside the pipeline module so the seam stays narrow.
+//!
+//! The legacy BM25 / fuzzy / inventory tools that used to share this
+//! file moved out in PR-G:
+//!   - `bm25_symbol_search` → `symbols/bm25_search.rs`
+//!   - `search_symbols_fuzzy` → `symbols/fuzzy_search.rs`
+//!   - `refresh_symbol_index` / `get_complexity` / `get_project_structure`
+//!     → `symbols/inventory.rs`
+
+use super::super::AppState;
+use crate::tool_runtime::ToolResult;
 use crate::tools::symbol_query::{SymbolQueryPipeline, SymbolQueryRequest};
-use codelens_engine::{SymbolKind, read_file, search_symbols_hybrid_with_semantic};
-use serde_json::{Value, json};
+use serde_json::Value;
 
 pub fn get_symbols_overview(state: &AppState, arguments: &Value) -> ToolResult {
-    crate::tools::symbol_query::SymbolQueryPipeline::new(state)
-        .run(crate::tools::symbol_query::SymbolQueryRequest::SymbolsOverview { arguments })
+    SymbolQueryPipeline::new(state).run(SymbolQueryRequest::SymbolsOverview { arguments })
 }
 
 pub fn find_symbol(state: &AppState, arguments: &Value) -> ToolResult {
-    crate::tools::symbol_query::SymbolQueryPipeline::new(state)
-        .run(crate::tools::symbol_query::SymbolQueryRequest::FindSymbol { arguments })
-}
-
-pub fn bm25_symbol_search(state: &AppState, arguments: &Value) -> ToolResult {
-    let query = required_string(arguments, "query")?;
-    let query_analysis = analyze_retrieval_query(query);
-    let max_results = optional_usize(arguments, "max_results", 10);
-    let include_tests = optional_bool(arguments, "include_tests", false);
-    let include_generated = optional_bool(arguments, "include_generated", false);
-    let session = crate::session_context::SessionRequestContext::from_json(arguments);
-    let scored = sparse_symbol_hits_for_query(
-        state,
-        &query_analysis,
-        max_results,
-        include_tests,
-        include_generated,
-        &session,
-    )?;
-
-    let total_query_terms = unique_query_terms(&query_analysis.expanded_query).len();
-    let payload_results: Vec<Value> = scored
-        .into_iter()
-        .enumerate()
-        .map(|(idx, hit)| {
-            let follow_up = suggested_follow_up(&hit.document.kind, hit.document.exported);
-            let confidence = confidence_tier(
-                &hit.matched_terms,
-                total_query_terms,
-                &hit.document.name,
-                &hit.document.name_path,
-            );
-            json!({
-                "symbol_id": hit.document.symbol_id,
-                "name": hit.document.name,
-                "name_path": hit.document.name_path,
-                "kind": hit.document.kind,
-                "file_path": hit.document.file_path,
-                "module_path": hit.document.module_path,
-                "signature": hit.document.signature,
-                "language": hit.document.language,
-                "line": hit.document.line_start,
-                "score": ((hit.score * 1000.0).round() / 1000.0),
-                "why_matched": hit.matched_terms,
-                "flags": {
-                    "is_test": hit.document.is_test,
-                    "is_generated": hit.document.is_generated,
-                    "exported": hit.document.exported,
-                },
-                "provenance": {
-                    "source": "sparse_bm25f",
-                    "retrieval_rank": idx + 1,
-                },
-                "suggested_follow_up": follow_up,
-                "confidence": confidence,
-            })
-        })
-        .collect();
-
-    let query_type = if query_analysis.prefer_lexical_only {
-        "identifier"
-    } else if query_analysis.natural_language {
-        "natural_language"
-    } else {
-        "short_phrase"
-    };
-    let retrieval = json!({
-        "lane": "sparse_bm25f",
-        "query_type": query_type,
-        "recommended": query_analysis.prefer_sparse_symbol_search,
-        "lexical_query": query_analysis.expanded_query,
-        "semantic_query": query_analysis.semantic_query,
-    });
-    let meta = success_meta(BackendKind::Sqlite, 0.88);
-    let evidence = crate::tool_evidence::tool_evidence(
-        "retrieval",
-        &meta,
-        "sparse_bm25f",
-        json!({
-            "preferred_lane": "sparse_bm25f",
-            "query_type": query_type,
-            "semantic_enabled": false,
-            "semantic_used_in_core": false,
-            "sparse_used_in_core": true,
-            "semantic_evidence_count": 0,
-            "sparse_evidence_count": payload_results.len(),
-            "precise_available": false,
-            "precise_used": false,
-            "precise_source": null,
-            "fallback_source": "sparse_bm25f",
-            "precise_result_count": 0,
-        }),
-    );
-
-    Ok((
-        json!({
-            "query": query,
-            "results": payload_results,
-            "count": payload_results.len(),
-            "retrieval": retrieval,
-            "evidence": evidence,
-        }),
-        meta,
-    ))
+    SymbolQueryPipeline::new(state).run(SymbolQueryRequest::FindSymbol { arguments })
 }
 
 pub fn get_ranked_context(state: &AppState, arguments: &Value) -> ToolResult {
     SymbolQueryPipeline::new(state).run(SymbolQueryRequest::RankedContext { arguments })
-}
-
-pub fn refresh_symbol_index(state: &AppState, _arguments: &Value) -> ToolResult {
-    let stats = state.symbol_index().refresh_all()?;
-    state.graph_cache().invalidate();
-    #[cfg(feature = "semantic")]
-    let mut payload = json!(stats);
-    #[cfg(not(feature = "semantic"))]
-    let payload = json!(stats);
-    #[cfg(feature = "semantic")]
-    {
-        let project = state.project();
-        let guard = state.embedding_ref();
-        if let Some(engine) = guard.as_ref()
-            && engine.is_indexed()
-        {
-            match engine.ensure_index_fresh_for_project(&project) {
-                Ok(report) => {
-                    if let Some(map) = payload.as_object_mut() {
-                        map.insert("embedding_freshness".to_owned(), json!(report));
-                    }
-                }
-                Err(error) => {
-                    if let Some(map) = payload.as_object_mut() {
-                        map.insert(
-                            "embedding_freshness".to_owned(),
-                            json!({
-                                "status": "unavailable",
-                                "reason": error.to_string()
-                            }),
-                        );
-                    }
-                }
-            }
-        }
-    }
-    Ok((payload, success_meta(BackendKind::TreeSitter, 0.95)))
-}
-
-pub fn get_complexity(state: &AppState, arguments: &Value) -> ToolResult {
-    let path = required_string(arguments, "path")?;
-    let symbol_name = optional_string(arguments, "symbol_name");
-    let file_result = read_file(&state.project(), path, None, None)?;
-    let lines = file_result.content.lines().collect::<Vec<_>>();
-    let symbols = state.symbol_index().get_symbols_overview_cached(path, 2)?;
-
-    let functions = flatten_symbols(&symbols)
-        .into_iter()
-        .filter(|s| matches!(s.kind, SymbolKind::Function | SymbolKind::Method))
-        .filter(|s| symbol_name.is_none_or(|name| s.name == name))
-        .map(|s| {
-            let start = s.line.saturating_sub(1).min(lines.len());
-            let end = (s.line + 50).min(lines.len());
-            let branches = count_branches(&lines[start..end]);
-            json!({
-                "name": s.name,
-                "kind": s.kind.as_label(),
-                "file": s.file_path,
-                "line": s.line,
-                "branches": branches,
-                "complexity": 1 + branches
-            })
-        })
-        .collect::<Vec<_>>();
-
-    let results = if functions.is_empty() {
-        let branches = count_branches(&lines);
-        vec![json!({
-            "name": path,
-            "branches": branches,
-            "complexity": 1 + branches
-        })]
-    } else {
-        functions
-    };
-
-    let avg_complexity = if results.is_empty() {
-        0.0
-    } else {
-        results
-            .iter()
-            .filter_map(|e| e.get("complexity").and_then(|v| v.as_i64()))
-            .map(|v| v as f64)
-            .sum::<f64>()
-            / results.len() as f64
-    };
-
-    Ok((
-        json!({
-            "path": path,
-            "functions": results,
-            "count": results.len(),
-            "avg_complexity": avg_complexity
-        }),
-        success_meta(BackendKind::TreeSitter, 0.89),
-    ))
-}
-
-#[allow(dead_code)]
-pub fn get_project_structure(state: &AppState, _arguments: &Value) -> ToolResult {
-    let dirs = state.symbol_index().get_project_structure()?;
-    let total_files: usize = dirs.iter().map(|d| d.files).sum();
-    let total_symbols: usize = dirs.iter().map(|d| d.symbols).sum();
-    Ok((
-        json!({
-            "directories": dirs,
-            "total_files": total_files,
-            "total_symbols": total_symbols,
-            "dir_count": dirs.len()
-        }),
-        success_meta(BackendKind::Sqlite, 0.95),
-    ))
-}
-
-pub fn search_symbols_fuzzy(state: &AppState, arguments: &Value) -> ToolResult {
-    let query = required_string(arguments, "query")?;
-    let max_results = optional_usize(arguments, "max_results", 30);
-    let fuzzy_threshold = arguments
-        .get("fuzzy_threshold")
-        .and_then(|v| v.as_f64())
-        .unwrap_or(0.6);
-    let disable_semantic = optional_bool(arguments, "disable_semantic", false);
-    // Build semantic scores if embeddings are available (same pattern as get_ranked_context)
-    let semantic_scores = semantic_scores_for_query(state, query, 50, disable_semantic);
-
-    let sem_ref = if semantic_scores.is_empty() {
-        None
-    } else {
-        Some(&semantic_scores)
-    };
-
-    let backend = if sem_ref.is_some() {
-        BackendKind::Hybrid
-    } else {
-        BackendKind::Sqlite
-    };
-
-    let pagerank_scores = state.graph_cache().file_pagerank_scores(&state.project());
-    let pagerank_ref = if pagerank_scores.is_empty() {
-        None
-    } else {
-        Some(pagerank_scores.as_ref())
-    };
-
-    Ok(search_symbols_hybrid_with_semantic(
-        &state.project(),
-        query,
-        max_results,
-        fuzzy_threshold,
-        sem_ref,
-        pagerank_ref,
-    )
-    .map(|value| {
-        (
-            json!({ "results": value, "count": value.len() }),
-            success_meta(backend, 0.9),
-        )
-    })?)
-}
-
-// ── Helpers ──────────────────────────────────────────────────────────────
-
-/// Follow-up tool hints for a BM25 symbol card.
-///
-/// Mirrors the `bm25-sparse-lane-spec` matrix. Frontier-model harnesses
-/// select their next tool off this list, so the output is part of the
-/// response contract. Keep it short (1-3 entries) — the goal is
-/// guidance, not an exhaustive menu.
-fn suggested_follow_up(kind: &str, exported: bool) -> Vec<&'static str> {
-    let base: Vec<&'static str> = match kind {
-        "function" | "method" => vec!["find_symbol", "get_file_diagnostics"],
-        "class" | "interface" | "enum" | "type_alias" => {
-            vec!["find_symbol", "find_referencing_symbols"]
-        }
-        "module" | "file" => vec!["get_symbols_overview", "find_referencing_symbols"],
-        "variable" | "property" => vec!["find_symbol", "find_referencing_symbols"],
-        _ => vec!["find_symbol"],
-    };
-    if exported
-        && matches!(kind, "function" | "method" | "class" | "interface")
-        && !base.contains(&"find_referencing_symbols")
-    {
-        let mut with_refs = base.clone();
-        with_refs.push("find_referencing_symbols");
-        return with_refs;
-    }
-    base
-}
-
-#[cfg(test)]
-mod suggested_follow_up_tests {
-    use super::suggested_follow_up;
-
-    #[test]
-    fn function_gets_body_then_diagnostics() {
-        let hints = suggested_follow_up("function", false);
-        assert_eq!(hints.first().copied(), Some("find_symbol"));
-        assert!(hints.contains(&"get_file_diagnostics"));
-    }
-
-    #[test]
-    fn class_gets_body_and_references() {
-        let hints = suggested_follow_up("class", false);
-        assert_eq!(hints, vec!["find_symbol", "find_referencing_symbols"]);
-    }
-
-    #[test]
-    fn module_gets_overview_first() {
-        let hints = suggested_follow_up("module", false);
-        assert_eq!(hints.first().copied(), Some("get_symbols_overview"));
-    }
-
-    #[test]
-    fn exported_function_also_offers_references() {
-        let hints = suggested_follow_up("function", true);
-        assert!(hints.contains(&"find_referencing_symbols"));
-        assert!(hints.contains(&"find_symbol"));
-    }
-
-    #[test]
-    fn unknown_kind_falls_back_to_find_symbol() {
-        let hints = suggested_follow_up("unknown", false);
-        assert_eq!(hints, vec!["find_symbol"]);
-    }
-}
-
-/// Cross-field confidence tier for a BM25 symbol card.
-///
-/// Without a separate dense arm, we cannot yet compute a true
-/// BM25-vs-dense agreement signal. This heuristic is the *cross-field*
-/// proxy: a result that matches query terms on the high-weight
-/// identifier fields (`name`, `name_path`) **and** covers most of the
-/// unique query terms is a high-confidence hit; a result that matches
-/// only on low-weight fields (body lexical chunk, doc comment) is low.
-///
-/// - `high`   — ≥80% query-term coverage AND a hit on name or name_path
-/// - `medium` — 2+ matched terms OR a name/name_path hit
-/// - `low`    — single term hit, or matches only on body/doc fields
-///
-/// Frontier-model callers use this to decide whether to trust the card
-/// for direct consumption or to cross-check via `find_symbol` +
-/// `find_referencing_symbols` before acting.
-fn confidence_tier(
-    matched_terms: &[String],
-    unique_query_terms: usize,
-    name: &str,
-    name_path: &str,
-) -> &'static str {
-    if matched_terms.is_empty() || unique_query_terms == 0 {
-        return "low";
-    }
-    let coverage = matched_terms.len() as f64 / unique_query_terms as f64;
-    let name_lower = name.to_ascii_lowercase();
-    let name_path_lower = name_path.to_ascii_lowercase();
-    let identifier_hit = matched_terms.iter().any(|term| {
-        let term_lower = term.to_ascii_lowercase();
-        name_lower.contains(&term_lower) || name_path_lower.contains(&term_lower)
-    });
-
-    if coverage >= 0.8 && identifier_hit {
-        "high"
-    } else if identifier_hit || matched_terms.len() >= 2 {
-        "medium"
-    } else {
-        "low"
-    }
-}
-
-#[cfg(test)]
-mod confidence_tier_tests {
-    use super::confidence_tier;
-
-    #[test]
-    fn full_coverage_on_name_path_is_high() {
-        let matched = vec!["dispatch".to_owned(), "tool".to_owned()];
-        assert_eq!(
-            confidence_tier(&matched, 2, "dispatch_tool", "dispatch::dispatch_tool"),
-            "high"
-        );
-    }
-
-    #[test]
-    fn partial_coverage_with_name_hit_is_medium() {
-        let matched = vec!["dispatch".to_owned()];
-        assert_eq!(
-            confidence_tier(&matched, 3, "dispatch_tool", "dispatch::dispatch_tool"),
-            "medium"
-        );
-    }
-
-    #[test]
-    fn body_only_match_is_low() {
-        let matched = vec!["invoke".to_owned()];
-        assert_eq!(
-            confidence_tier(&matched, 2, "dispatch_tool", "dispatch::dispatch_tool"),
-            "low"
-        );
-    }
-
-    #[test]
-    fn multiple_matches_without_name_hit_is_medium() {
-        let matched = vec!["invoke".to_owned(), "handler".to_owned()];
-        assert_eq!(
-            confidence_tier(&matched, 3, "dispatch_tool", "dispatch::dispatch_tool"),
-            "medium"
-        );
-    }
-
-    #[test]
-    fn empty_matched_is_low() {
-        assert_eq!(confidence_tier(&[], 2, "x", "a::x"), "low");
-    }
-
-    #[test]
-    fn zero_query_terms_is_low() {
-        let matched = vec!["dispatch".to_owned()];
-        assert_eq!(
-            confidence_tier(&matched, 0, "dispatch_tool", "dispatch::dispatch_tool"),
-            "low"
-        );
-    }
 }
